Move berttokenizer
