+++ capreolus/tokenizer/__init__.py
@@ -1,4 +1,5 @@
 from capreolus.registry import ModuleBase, RegisterableModule, Dependency
+from transformers import BertTokenizer as HFBertTokenizer
 
 
 class Tokenizer(ModuleBase, metaclass=RegisterableModule):
@@ -41,3 +42,26 @@ class AnseriniTokenizer(Tokenizer):
             return self._tokenize(sentences)
 
         return [self._tokenize(s) for s in sentences]
+
+class BertTokenizer(Tokenizer):
+    name = "berttokenizer"
+
+    @staticmethod
+    def config():
+        pass
+
+    def __init__(self, cfg):
+        super().__init__(cfg)
+        self.bert_tokenizer = HFBertTokenizer.from_pretrained(cfg["pretrained"])
+
+    def convert_tokens_to_ids(self, tokens):
+        return self.bert_tokenizer.convert_tokens_to_ids(tokens)
+
+    def tokenize(self, sentences):
+        if not sentences or len(sentences) == 0:  # either "" or []
+            return []
+
+        if isinstance(sentences, str):
+            return self.bert_tokenizer.tokenize(sentences)
+
+        return [self.bert_tokenizer.tokenize(s) for s in sentences]
